# Conflicts:
#	src/main/java/module-info.java
#	src/main/resources/initsql.sql

+++ src/main/java/it/unipv/ingsw/lasout/facade/cashbook/CashbookFacade.java
@@ -4,12 +4,14 @@ import it.unipv.ingsw.lasout.model.cashbook.Cashbook;
 import it.unipv.ingsw.lasout.model.cashbook.ICashbookDAO;
 import it.unipv.ingsw.lasout.model.cashbook.exception.CannotDeleteDefaultCashbookException;
 import it.unipv.ingsw.lasout.model.cashbook.exception.CashbookAlreadyExistingException;
+import it.unipv.ingsw.lasout.model.transaction.AutomaticTransaction;
 import it.unipv.ingsw.lasout.model.transaction.ModifiableTransaction;
 import it.unipv.ingsw.lasout.model.transaction.Transaction;
 import it.unipv.ingsw.lasout.model.transaction.exception.CannotEditTransactionException;
 import it.unipv.ingsw.lasout.model.user.User;
 import it.unipv.ingsw.lasout.util.DaoFactory;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class CashbookFacade implements ICashbookFacade {
@@ -61,19 +63,31 @@ public class CashbookFacade implements ICashbookFacade {
     @Override
     public boolean deleteCashbook(Cashbook cashbook){
         try{
-            cashBookDAO.delete(cashbook);
-        } catch (CannotDeleteDefaultCashbookException e) {
-            throw new CannotDeleteDefaultCashbookException(e.getMessage());
+            //verifico non sia un cashbook default
+            if(!cashBookDAO.getRaw(cashbook).isDefault()){
+                cashBookDAO.delete(cashbook);
+                return true;
+            } else {
+                throw new CannotDeleteDefaultCashbookException();
+            }
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
-        return true;
     }
 
     @Override
     public List<Cashbook> getUserCashbooks(User carrierUser){
         try{
             return cashBookDAO.getAllUserCashbooks(carrierUser);
+        } catch (Exception e) {
+            return null;
+        }
+    }
+
+    @Override
+    public Cashbook getUserDefaultCashbook(User carrierUser){
+        try{
+            return cashBookDAO.getDefaultCashbook(carrierUser);
         } catch (Exception e) {
             return null;
         }
@@ -94,10 +108,11 @@ public class CashbookFacade implements ICashbookFacade {
 
     /**
      * Serve a salvare una transazione conoscendo lo user che la ha eseguita
-     * @param user
-     * @param transaction
-     * @return
+     * @param user of which you want to add the transaction
+     * @param transaction transaction to add
+     * @return true if transaction was added correctly, false otherwise
      */
+    @Override
     public boolean addTransaction(User user, Transaction transaction) {
         Cashbook defaultCashbook = cashBookDAO.getDefaultCashbook(user);
         return addTransaction(defaultCashbook, transaction);
@@ -112,7 +127,7 @@ public class CashbookFacade implements ICashbookFacade {
             cashBookDAO.update(c);
             editCashbook(c);
         } catch (CannotEditTransactionException e) {
-            return false;
+            throw e;
         } catch (Exception e) {
             return false;
         }
@@ -131,6 +146,45 @@ public class CashbookFacade implements ICashbookFacade {
             return false;
         }
         return true;
+    }
+
+    @Override
+    public double calculateSummary(Cashbook cashbook){
+        double sum = 0;
+        try {
+            List<Transaction> transactionList = cashBookDAO.get(cashbook).getTransactionList();
+            if (transactionList != null) {
+                for (Transaction transaction : transactionList) {
+                    sum += transaction.getAmount();
+                }
+            }
+            return sum;
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public List<Transaction> getAutomaticTransactions(Cashbook c){
+        List<Transaction> transactionList = c.getTransactionList();
+        List<Transaction> automaticTransactions = new ArrayList<Transaction>();
+
+        for(Transaction transaction : transactionList){
+            try{
+                AutomaticTransaction automaticTransaction = (AutomaticTransaction) transaction;
+                automaticTransactions.add(automaticTransaction);
+            } catch (Exception e) {
+                continue;
+            }
+        }
+
+        return automaticTransactions;
+    }
+
+    @Override
+    public List<Transaction> getVaultTransactionsOfUser(User user){
+        Cashbook c = getUserDefaultCashbook(user);
+        return getAutomaticTransactions(c);
     }
 
 }
